refactor: Move around more stuff to prevent circular dependencies
