+++ src/aux-common/Files/FileCalculationContext.ts
