+++ src/aux-common/Files/index.ts
@@ -1,5 +1,8 @@
 export * from './File';
-export * from './FileContext';
+export * from './FileEvents';
+export * from './FileCalculationContext';
 export * from './FileCalculations';
+export * from './FileCalculationContextFactories';
 export * from './FilesChannel';
+export * from './FileActions';
 export * from './TagUpdatedEvent';
